PRO-4814: fix failing pact tests
