--- src/contractTest/resources/json/intestacyGrantOfRepresentation_invalid_presubmit_response.json
@@ -1,1 +1,0 @@
-
